Resolved a bug related to sphere shield
The issue was resolved by disabling the script, collder and networkObject scripts on the sphere instantiated locally.

Spawning a sphere shield on the client works well and is destroyed by the host correctly. Spawning one on the host was causing an error due to the way spawning a shield is handled: instantiate and spawning seperately.
I believe that the issue only replicated when the sphere shield was spawning first on the client and then on the host. Either that or it was a case of which shield the projectile was hitting first.

- Other minor changes made

- WIP finalizing the projectile script to work with new architecture.