Fixed: Added local parry before despawn
There was an issue parrying the projectile. The suspicion is that due to the fact that the parry was using a serverRPC to despawn the object, the delay caused the parry action to not be immediate nor responsive.

To that end, a local ActiveSelf(false) was added to the projectile to first deactivate it before it is despawned.